requirements.txt
openai
fastapi
uvicorn
psycopg2-binary
pgvector
python-dotenv
pydantic
pandas
requests
